+++ Server/Adapters/TerrariaBuildContext.cs
@@ -4,25 +4,25 @@ using PacketManager.Core.Data;
 namespace PacketManager.Server.Adapters;
 
 /// <summary>
-/// Реализация контекста сборки пакета для Terraria.
-/// Расширяет стандартный контекст дополнительным полем Num, 
-/// получаемым из оригинального метода OnPacketWrite.
+/// Реализация контекста сборки пакета для <see cref="Terraria"/>.
+/// Расширяет стандартный контекст дополнительным полем <see cref="num"/>, 
+/// получаемым из оригинального метода <see cref="Terraria.NetMessage.OnPacketWrite"/>.
 /// </summary>
 /// <remarks>
-/// Создает новый экземпляр контекста для Terraria.
+/// Создает новый экземпляр контекста для <see cref="Terraria"/>.
 /// </remarks>
 /// <param name="messageId">Идентификатор типа пакета.</param>
 /// <param name="writer">Бинарный писатель.</param>
-/// <param name="targets">Список клиентов Terraria.</param>
+/// <param name="targets">Список клиентов <see cref="Terraria"/>.</param>
 /// <param name="data">Оригинальные данные пакета.</param>
-/// <param name="num">Вспомогательное числовое значение из OnPacketWrite.</param>
-internal class TerrariaBuildContext(byte messageId, BinaryWriter writer,
+/// <param name="num">Вспомогательное числовое значение из <see cref="Terraria.NetMessage.OnPacketWrite"/>.</param>
+internal class TerrariaBuildContext(int messageId, BinaryWriter writer,
     List<TerrariaNetworkClient> targets, PacketData data, int num) : IPacketBuildContext
 {
     /// <summary>
     /// Получает идентификатор типа пакета.
     /// </summary>
-    public byte MessageId { get; } = messageId;
+    public int MessageId { get; } = messageId;
 
     /// <summary>
     /// Получает бинарный писатель для записи данных.
@@ -30,7 +30,7 @@ internal class TerrariaBuildContext(byte messageId, BinaryWriter writer,
     public BinaryWriter Writer { get; } = writer;
 
     /// <summary>
-    /// Получает список целевых клиентов Terraria.
+    /// Получает список целевых клиентов <see cref="Terraria"/>.
     /// </summary>
     public IReadOnlyCollection<INetworkClient> Targets { get; } = targets;
 
@@ -40,8 +40,8 @@ internal class TerrariaBuildContext(byte messageId, BinaryWriter writer,
     public PacketData OriginalData { get; } = data;
 
     /// <summary>
-    /// Получает вспомогательное числовое значение Num, переданное из OnPacketWrite.
-    /// Может использоваться для специфичной логики Terraria (например, индекса игрока).
+    /// Получает вспомогательное числовое значение <see cref="num"/>, переданное из <see cref="Terraria.NetMessage.OnPacketWrite"/>.
+    /// Может использоваться для специфичной логики <see cref="Terraria"/> (например, индекса игрока).
     /// </summary>
     public int Num { get; } = num;
 }
